before refactor

+++ gui/app/models/base_file_model.rb
@@ -28,26 +28,30 @@ class BaseFileModel
   # Get json from a file
   #-----------------------------------------
   def self.get_json_data(file_path)
-    file1 = File.open(file_path)
-    text = file1.read
-    file1.close
-    
-    # wipe text
-    #text=text.grep(/^\s*[^#]/).to_s
-    text=text.gsub(/^\s*#.*$/,'').gsub(/^\n$/,'')
+	
+	  data=nil
+				
+	  if File.exists?(file_path)
+				file1 = File.open(file_path)
+				text = file1.read
+				file1.close
+				
+				# wipe text
+				#text=text.grep(/^\s*[^#]/).to_s
+				text=text.gsub(/^\s*#.*$/,'').gsub(/^\n$/,'')
 
-    # decode json
-    # data = ActiveSupport::JSON.decode(text)
-    data=nil
-    # decode json
-    if !text.blank?
-    
-    #schema = Kwalify::Yaml.load_file('schema.yaml')
-    #validator = Kwalify::Validator.new(schema)
-	    data = ActiveSupport::JSON.decode(text)
-	  end
-    #data = JSON.parse(text)
-
+				# decode json
+				# data = ActiveSupport::JSON.decode(text)
+				# decode json
+				if !text.blank?
+				
+				#schema = Kwalify::Yaml.load_file('schema.yaml')
+				#validator = Kwalify::Validator.new(schema)
+					data = ActiveSupport::JSON.decode(text)
+				end
+				#data = JSON.parse(text)
+		end
+		
     return data
   end
